test: fix ci error(subprocess killed was still pending)

+++ packages/dever/__tests__/index.spec.ts
@@ -31,14 +31,21 @@ describe('index', () => {
 			stdio: 'pipe'
 		});
 		let stdout = '';
-		subprocess.stdout.on('data', (data) => {
-			stdout += data.toString().replace(/(\t|\n|\v|\r|\f|\s)/g, '');
-			if (!stdout.includes('index.html')) return;
-			subprocess.kill();
+		const ready = new Promise<void>(resolve => {
+			subprocess.stdout.on('data', (data) => {
+				stdout += data.toString().replace(/(\t|\n|\v|\r|\f|\s)/g, '');
+				if (!stdout.includes('index.html')) return;
+				resolve();
+			});
 		});
+		
 		try {
-			await subprocess;
-		} catch { /* empty */ }
+			await ready;
+			subprocess.kill();
+		} catch (e) { 
+			process.env.CI && console.log(e);
+			!subprocess.killed && subprocess.kill();
+		}
 		expect(stdout).toMatch(/>index:http.*\/index\.html/);
 	}, 60000);
 
@@ -50,60 +57,67 @@ describe('index', () => {
 			stdio: 'pipe'
 		});
 		let expects = ['/components/button/index.html', '/vue/index.html', '/react/index.html'];
-		const run = (url: string, timeout: number) => {
-			let ctx: Launch;
-			Promise.resolve()
-				.then(() => {
-					ctx = new Launch();
-					return ctx.createPage();
-				})
-				.then(() => {
-					return ctx.page.goto(url, { timeout });
-				})
-				.then(() => {
-					return Promise.all([
-						ctx.operater.html("#test"),
-						ctx.operater.classList("#app")
-					]);
-				})
-				.then(([html, classList]) => {
-					expect(html).toMatch('Hello World!');
-					expect(classList).toEqual(['preload']);
-					expects = expects.filter(i => !url.includes(i));
-					if (!expects.length) {
-						ctx.browser.close().finally(() => {
-							subprocess.kill();
-						});
+
+		const ready = new Promise<void>(resolve => {
+			const run = (url: string, timeout: number) => {
+				let ctx: Launch;
+				Promise.resolve()
+					.then(() => {
+						ctx = new Launch();
+						return ctx.createPage();
+					})
+					.then(() => {
+						return ctx.page.goto(url, { timeout });
+					})
+					.then(() => {
+						return Promise.all([
+							ctx.operater.html("#test"),
+							ctx.operater.classList("#app")
+						]);
+					})
+					.then(([html, classList]) => {
+						expect(html).toMatch('Hello World!');
+						expect(classList).toEqual(['preload']);
+						expects = expects.filter(i => !url.includes(i));
+
+						if (!expects.length) {
+							ctx.browser.close().finally(resolve);
+						}
+					})
+					.catch((e) => {
+						ctx.browser.close()
+							.then(
+								() => {
+									if (String(e).includes('TimeoutError')) {
+										process.env.CI && console.log(url, timeout, String(e).replace(/(.*)\n.*/, '$1'), /try again/);
+										run(url, timeout + 100);
+									} else {
+										console.log(e, /task error/);
+										throw e;		
+									}
+								},
+								resolve	
+							);
+					});
+			};
+			subprocess.stdout.on('data', (data) => {
+				data = data.toString().replace(/(\t|\n|\v|\r|\f|\s)/g, '');
+				if (!data) return;
+				data.split('➜')[0].split('>').filter((i: any) => !!i).forEach((url: string) => {
+					url = url.match(/(.*)(http:.*)/)?.[2] || '';
+					if (url && expects.some(i => url.includes(i))) {
+						run(url, 500);
 					}
-				})
-				.catch((e) => {
-					ctx.browser.close()
-						.then(
-							() => {
-								if (String(e).includes('TimeoutError')) {
-									process.env.CI && console.log(url, String(e).replace(/(.*)\n.*/, '$1'), /try again/);
-									run(url, timeout + 100);
-								} else {
-									console.log(e, /task error/);
-									throw e;		
-								}
-							},
-							() => subprocess.kill()		
-						);
 				});
-		};
-		subprocess.stdout.on('data', (data) => {
-			data = data.toString().replace(/(\t|\n|\v|\r|\f|\s)/g, '');
-			if (!data) return;
-			data.split('➜')[0].split('>').filter((i: any) => !!i).forEach((url: string) => {
-				url = url.match(/(.*)(http:.*)/)?.[2] || '';
-				if (url && expects.some(i => url.includes(i))) {
-					run(url, 500);
-				}
 			});
-		}); 
+		});
+		 
 		try {
-			await subprocess;
-		} catch { /* empty */ }
+			await ready;
+			subprocess.kill();
+		} catch (e) { 
+			process.env.CI && console.log(e);
+			!subprocess.killed && subprocess.kill();
+		}
 	}, 150000);
 });
